requirements.txt
httpx
beautifulsoup4
html2text
openai
python-dotenv
loguru
fastapi
uvicorn
